refactor(menu):merge menu-sub to menu
