+++ src/app/live-demo/menu/app.component.html
@@ -4,13 +4,13 @@
         <div class="demo-1 live-demo live-demo-block live-demo-menu">
             <h3>基本</h3>
             <p class="comment">基本用法</p>
-            <j-menu [data]="titles" [width]="'160'" (select)="selectChange($event)"></j-menu>
+            <j-menu [data]="titles" [width]="'160'" (select)="selectChange($event)" root></j-menu>
         </div>
 
         <div class="demo-2 live-demo live-demo-block live-demo-menu">
             <h3></h3>
             <p class="comment"></p>
-            <j-menu #jMenu [width]="'300'"></j-menu>
+            <j-menu #jMenu [width]="'300'" root></j-menu>
         </div>
 
         <div class="demo-2 live-demo live-demo-block live-demo-menu">
