--- src/jigsaw/component/menu/menu-sub.component.scss
@@ -1,16 +1,0 @@
-.jigsaw{
-    &-menu-parent{
-        position: relative;
-    }
-    &-menu-child{
-        position: absolute;
-    }
-    &-menu-hide{
-        display: none;
-    }
-    &-menu-icon{
-        display: inline-block;
-        font-size: inherit;
-        margin-right: 1em;
-    }
-}
