+++ src/jigsaw/component/menu/menu.component.html
@@ -1,4 +1,12 @@
-<nav>
-    <j-menu-sub [width]="width" [height]="height" [item]="_titles" root></j-menu-sub>
-</nav>
-
+<j-list [width]="width" [height]="height" trackItemBy="label" [ngClass]="{'jigsaw-menu-child jigsaw-menu-hide': !_root}" (selectedItemsChange)="_handleSelect($event)">
+    <ng-template ngFor let-child [ngForOf]="_items">
+        <j-list-option [value]="child" [ngClass]="{'jigsaw-menu-parent':child.children}" (mouseenter)="_onListMouseEnter($event, child)" (mouseleave)="_onListMouseLeave($event)">
+            <div class="jigsaw-menu-icon" j-title>
+                <i [class]="child.icon"></i>
+            </div>
+            <span j-title>{{child.label}}</span>
+            <div j-sub-title>{{child.extraLabel}} <span *ngIf="child.children" class="fa fa-angle-right"></span></div>
+            <j-menu [width]="_width" [height]="_height" [data]="child.children" *ngIf="child.children"></j-menu>
+        </j-list-option>
+    </ng-template>
+</j-list>
